pyproject.toml
[tool.poetry]
name = "testtask"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
pandas = "^2.2.1"
palettable = "^3.3.3"
lxml = "^5.1.0"
html5lib = "^1.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"
django-bootstrap5 = "^23.4"
django-bootstrap-datepicker-plus = "^5.0.5"
django-select2 = "^8.1.2"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
